Started to refactor ClassroomModal by breaking each component into its own file. Creates ClassroomModal folder to later contain each component and util files for ClassroomModal.
